+++ Structs/CharaView/CharaView.cs
@@ -25,7 +25,7 @@ public unsafe partial struct CharaView : ICreatable
     [FieldOffset(0x14)] public uint CameraType; // turns portrait ambient/directional lighting on/off
     [FieldOffset(0x18)] public nint CameraManager;
     [FieldOffset(0x20)] public Camera* Camera;
-    //[FieldOffset(0x28)] public nint Unk28;
+    //[FieldOffset(0x28)] public nint Unk28; // float CharacterRotation?
     [FieldOffset(0x30)] public AgentInterface* Agent; // for example: AgentTryOn
     //[FieldOffset(0x38)] public nint AgentCallbackReady; // if set, called when State changes to Ready
     //[FieldOffset(0x40)] public nint AgentCallback; // not investigated, used inside vf7 and vf11
@@ -106,7 +106,7 @@ public unsafe partial struct CharaView : ICreatable
     public readonly partial void UnequipGear(bool hasCharacterData = false, bool characterLoaded = true);
 
     [MemberFunction("E8 ?? ?? ?? ?? 45 33 DB FF C3")]
-    public readonly partial void SetItemSlotData(byte slotId, uint itemId, byte stain, int a5 = 0, byte a6 = 1); // maybe a5 is glamour id and a6 is a boolean that it should use glamour?
+    public readonly partial void SetItemSlotData(byte slotId, uint itemId, byte stain, int a5 = 0, byte a6 = 1);
 
     [MemberFunction("E8 ?? ?? ?? ?? B1 01 0F B6 86")]
     public readonly partial void ToggleDrawWeapon(bool drawn);