+++ CHANGELOG.md
@@ -4,7 +4,21 @@
 
 ### New Tweak: Background Music Keybind
 
-The game already has a keybind to mute/unmute sounds in general, but does not have a separate keybind to toggle background music. This tweak allows you to configure such a keybind.
+Adds a configurable keybind to toggle the background music, in addition to the game's existing keybind option to toggle the sound. No more `/bgm` macro. Works in cutscenes.
+
+### New Tweak: Gear Set Grid
+
+A window that displays a grid view of all the items in each gear set, making it easier to figure out which items to upgrade next.
+
+Configuration options:
+
+- Auto-open/close with Gear Set List
+- Register `/gsg` command to toggle window
+- Allow switching gearsets  
+  Makes the id column clickable.
+- Convert separator gear set with spacing  
+  When using separator gear sets (e.g. a gearset with name ===========) this option automatically converts it into spacing between rows (in the Gear Set Grid).
+  - Spacing between rows can be fully eliminated with the "Disable spacing" option.
 
 ### Commands